decouple fake settings from service spec
- scan through all extensions and make minor improvements
- move self containing extension methods to corresponding feature /
  layer extension class
