+++ src/core/Baked.Architecture/Testing/Extensions/AssertionExtensions.cs
@@ -7,7 +7,10 @@ namespace Baked;
 public static class AssertionExtensions
 {
     [DoesNotReturn]
-    public static void ShouldFail(this Spec _, string message = "") => throw new AssertionException(message);
+    public static void ShouldFail(this Spec _, string message = "") =>
+        throw new AssertionException(message);
+
     [DoesNotReturn]
-    public static void ShouldPass(this Spec _, string message = "") => Assert.Pass(message);
+    public static void ShouldPass(this Spec _, string message = "") =>
+        Assert.Pass(message);
 }
