+++ src/recipe/Baked.Recipe.Service.Application/Communication/Http/HttpCommunicationExtensions.cs
@@ -7,4 +7,20 @@ public static class HttpCommunicationExtensions
 {
     public static HttpCommunicationFeature Http(this CommunicationConfigurator _) =>
         new();
+
+    internal static Dictionary<string, string> Merge(this Dictionary<string, string>? dictionary, Dictionary<string, string>? input)
+    {
+        dictionary ??= [];
+        input ??= [];
+
+        foreach (var (key, value) in input)
+        {
+            if (!dictionary.ContainsKey(key))
+            {
+                dictionary[key] = value;
+            }
+        }
+
+        return dictionary;
+    }
 }
