+++ src/recipe/Baked.Recipe.Service.Application/Core/Mock/FakeSettings.cs
@@ -1,0 +1,4 @@
+﻿namespace Baked.Core.Mock;
+
+public class FakeSettings : Dictionary<string, string>
+{ }
