+++ src/recipe/Baked.Recipe.Service.Application/Core/Mock/MockCoreFeature.cs
@@ -11,6 +11,7 @@ public class MockCoreFeature : IFeature<CoreConfigurator>
         configurator.ConfigureServiceCollection(services =>
         {
             services.AddSingleton<TimeProvider, ResettableFakeTimeProvider>();
+            services.AddSingleton<FakeSettings>();
         });
 
         configurator.ConfigureTestConfiguration(test =>
