+++ src/recipe/Baked.Recipe.Service.Application/Testing/TestingExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using Baked.Architecture;
 using Baked.Testing;
-using Microsoft.Extensions.Configuration;
 using Moq;
 using Moq.Language.Flow;
 
@@ -56,54 +55,5 @@ public static class TestingExtensions
 
             return results[currentResultIndex++];
         });
-    }
-
-    public static void ASetting<T>(this Mocker mockMe,
-        string? key = default,
-        T? value = default
-    ) => mockMe.ASetting(key: key, value: $"{value}");
-
-    public static void ASetting(this Mocker mockMe,
-        string? key = default,
-        string? value = default
-    )
-    {
-        key ??= "Test:Configuration";
-        value ??= "value";
-
-        var spec = (ServiceSpec)mockMe.Spec;
-
-        spec.Settings[key] = value;
-    }
-
-    public static IConfiguration TheConfiguration(this Mocker mockMe,
-        Func<string, string?>? defaultValueProvider = default,
-        Dictionary<string, string>? settings = default
-    )
-    {
-        defaultValueProvider ??= _ => default;
-        settings ??= [];
-
-        var configuration = mockMe.Spec.GiveMe.The<IConfiguration>();
-
-        Mock.Get(configuration)
-           .Setup(c => c.GetSection(It.IsAny<string>())).Returns((string key) =>
-           {
-               var mockSection = new Mock<IConfigurationSection>();
-
-               mockSection.Setup(s => s.Value).Returns(() =>
-               {
-                   if (settings.TryGetValue(key, out var result))
-                   {
-                       return result;
-                   }
-
-                   return defaultValueProvider(key);
-               });
-
-               return mockSection.Object;
-           });
-
-        return configuration;
     }
 }
